Act on item-links before any document is converted to docutils
The comment was right: adding the relationships from item-link
directives should be done **before** any object is converted to
docutils. The original location of the add_relation() call was indeed
before converting to docutils, but other documents were already
converted to docutils: so too late.

Now adding the relations directly when parsing the item-link (which is
consistent with parsing item directive.

note: it could be that at the time of parsing item-link, the item is not
defined yet. So the add_relation() function needs to allow undefined
source (same as undefined target) and create a placeholder when
needed...